fox32: Use the instruction size when reading from an immediate pointer
Previously, Operand::ImmediatePtr always read 32 bits from memory, then
truncated the value if the instruction size was smaller. Technically
this worked, but it would cause fox32 to print warnings if reading a
value at the very end of memory, because it would attempt to read past
the end of memory. Now, Operand::ImmediatePtr contains the instruction
size which is used by read_source().

This reverts commit 634a776a631c8e34f138dfbbfebf1c5cc74b6f05